refactor: 구구 피드백 적용
