+++ jdbc/src/main/java/nextstep/jdbc/JdbcTemplate.java
@@ -1,18 +1,24 @@
 package nextstep.jdbc;
 
+import java.sql.Connection;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.ArrayList;
 import java.util.List;
 import javax.sql.DataSource;
+import nextstep.jdbc.exception.DataAccessException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class JdbcTemplate extends AbstractJdbcTemplate{
+public class JdbcTemplate {
 
     private static final Logger log = LoggerFactory.getLogger(JdbcTemplate.class);
 
+    private DataSource dataSource;
+
     public JdbcTemplate(DataSource dataSource) {
-        super(dataSource);
+        this.dataSource = dataSource;
     }
 
     public void update(final String query, Object... args) {
@@ -40,6 +46,22 @@ public class JdbcTemplate extends AbstractJdbcTemplate{
         );
     }
 
+    public <T> T queryForObject(String sql, RowMapper<T> rowMapper, Object... args) {
+        return executeQuery(
+            (conn) -> {
+                PreparedStatement preparedStatement = conn.prepareStatement(sql);
+                setParameters(preparedStatement, args);
+                return preparedStatement;
+            },
+            (rs) -> {
+                if (rs.next()) {
+                    return rowMapper.mapRow(rs, 1);
+                }
+                return null;
+            }
+        );
+    }
+
     public <T> List<T> query(String sql, RowMapper<T> rowMapper) {
         return executeQuery(
             (conn) -> conn.prepareStatement(sql),
@@ -56,5 +78,49 @@ public class JdbcTemplate extends AbstractJdbcTemplate{
             },
             rowMapper
         );
+    }
+
+    protected void executeUpdate(StatementStrategy stmt) {
+        execute(stmt, PreparedStatement::executeUpdate);
+    }
+
+    private <T> T execute(StatementStrategy stmt, Executor<T> executor) {
+        try (Connection conn = dataSource.getConnection();
+            PreparedStatement ps = stmt.makePreparedStatement(conn);
+        ) {
+            return executor.doSomething(ps);
+        } catch (SQLException e) {
+            throw new DataAccessException(e.getMessage(), e);
+        }
+    }
+
+    protected <T> T executeQuery(StatementStrategy stmt, ResultSetExtractor<T> rse) {
+        return execute(stmt, (ps) -> {
+            try (ResultSet rs = ps.executeQuery()) {
+                return rse.extractData(rs);
+            }
+        });
+    }
+
+    protected <T> List<T> executeQuery(StatementStrategy stmt, RowMapper<T> rowMapper) {
+        try (Connection conn = dataSource.getConnection();
+            PreparedStatement ps = stmt.makePreparedStatement(conn);
+            ResultSet rs = ps.executeQuery();
+        ) {
+            List<T> result = new ArrayList<>();
+
+            int rowNum = 1;
+            while (rs.next()) {
+                result.add(rowMapper.mapRow(rs, rowNum));
+                rowNum++;
+            }
+            return result;
+        } catch (SQLException e) {
+            throw new DataAccessException(e.getMessage(), e);
+        }
+    }
+
+    private interface Executor<T> {
+        T doSomething(PreparedStatement ps) throws SQLException;
     }
 }
